tryLock update
git-svn-id: svn://svn.caucho.com/home/svn/svnroot/resin/trunk@10386 9c94448d-38f1-0310-a231-d98308ff1ebf

+++ modules/resin/src/com/caucho/db/sql/DeleteQuery.java
@@ -78,7 +78,9 @@ class DeleteQuery extends Query {
       }
 
       do {
-        context.lock();
+        if (! context.lock()) {
+          throw new IllegalStateException("unable to lock for delete");
+        }
         
         try {
           if (isSelect(context) && rows[0].delete()) {
